docs: add sankey component documentation
